requirements.txt
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
scikit-learn>=1.3.0
fredapi>=0.5.1
alpha-vantage>=2.3.1
requests>=2.31.0
jupyter>=1.0.0
notebook>=6.5.0
python-dotenv>=1.0.0
streamlit>=1.28.0
plotly>=5.17.0

